Added support to configure USB network via file /etc/default/usbnetwork, by default all is disabled

+++ src/osso-usb-mass-storage-enable.sh
@@ -45,19 +45,27 @@ if [ $? = 0 ]; then
     else
         logger "$0: syncd is not running"
     fi
-    if [ -f /var/run/dnsmasq.pid.usb0 ]; then
-        DNSMASQ_PID=`cat /var/run/dnsmasq.pid.usb0`
-        rm -f /var/run/dnsmasq.pid.usb0
-        kill $DNSMASQ_PID
-    else
-        logger "$0: dnsmasq for USB network is not running"
+
+    [ -f /etc/default/usbnetwork ] && . /etc/default/usbnetwork
+    if [ "$USBNETWORK_ENABLE" = "1" ]; then
+        if [ "$USBNETWORK_DNSMASQ" = "1" ]; then
+            if [ -f /var/run/dnsmasq.pid.usb0 ]; then
+                DNSMASQ_PID=`cat /var/run/dnsmasq.pid.usb0`
+                rm -f /var/run/dnsmasq.pid.usb0
+                kill $DNSMASQ_PID
+            else
+                logger "$0: dnsmasq for USB network is not running"
+            fi
+        fi
+        if [ "$USBNETWORK_NAT" = "1" ]; then
+            if [ -f /proc/sys/net/ipv4/ip_forward -a -x /usr/sbin/iptables ]; then
+                echo 0 > /proc/sys/net/ipv4/ip_forward
+                /usr/sbin/iptables -t nat -D POSTROUTING ! -o lo -j MASQUERADE
+            fi
+        fi
+        rm -f /var/run/resolv.conf.usb0
+        ifdown usb0
     fi
-    if [ -f /proc/sys/net/ipv4/ip_forward -a -x /usr/sbin/iptables ]; then
-        echo 0 > /proc/sys/net/ipv4/ip_forward
-        /usr/sbin/iptables -t nat -D POSTROUTING ! -o lo -j MASQUERADE
-    fi
-    rm -f /var/run/resolv.conf.usb0
-    ifdown usb0
 
     sleep 2
     /sbin/rmmod g_nokia
